Refactored similar functions and such into their own packages.
This doesn't quite work like I want it to, but it is much more
organized. My main issue right now is symbol naming conflicts between
the :com.stone.solver-utils package and the :com.stone.nested-lists
package, as the former uses the latter, somehow causing this issue.
I probably need to read up more on packages so I understand them
fully, as well as look into using ASDF in order to facilitate the
loading of the packages (I'm not quite sure what it is either).