Made take-action handle nested-nth's restarts.
It now retuns nil if the given action isn't legal (at least if it
results in moving left or up off of the board). I still need to work
on implementing hard boundaries, however.

Also fixed take-actions-list so that it short-circuits to nil if any
of the supplied actions aren't legal on one of the intermediate states.