Left a few reminder comments for things to implement/remove later.
Added a reader for the previous-action slot in the puzzle-node class.

Removed unused generic function, as it was implemented with regular
functions. It also didn't make much sense to implement it as generic
anyways.

Renamed error-negative-index to describe its functionality more
accurately.

Removed the use-values restart from negative-index-restarts, as it is
already implemented via the builtin use-value.

Fixed some indentation and whitespace issues.