Added a node-take-best action function.
It returns the node associated with the best action along with
all of the other nodes for the purpose of adding to the frontier
once that is implemented.

Made legal-actions just return a list of actions that are legal
instead of the metadata associated with them.

Added a newline after printing a node's state in node-print-state.