The solver is not in a working state.
Exported necessary symbols from com.stone.solver-utils.

Created function net-node-score to take every tile's position and
the number of moves associated with the node into account for score
calculations.

Added *tile-choices* list for use in the aforementioned score function
as well as *minimal-goal-node* for comparison purposes.

Moved some functions around so as to not reference undefined
functions.

Added in documentation where it was missing.

Created two predicate functions for use in solving the puzzle based on
a node.

Fully implemented frontier/explored sets when solving the puzzle.

Added traverse-actions function as a way to list of all moves
necessary to solve the 15 puzzle.

Added initial-node-from-state function for easy creation of a starting
node.